Adding support for df['timestamp'].min() etc.
There is still a difference between pandas/eland in terms
of min/max etc. aggregations as pandas supports this
on strings.
