+++ eland/query_compiler.py
@@ -20,7 +20,7 @@ from typing import Union
 import numpy as np
 import pandas as pd
 
-from eland import Client, DEFAULT_PROGRESS_REPORTING_NUM_ROWS
+from eland import Client, DEFAULT_PROGRESS_REPORTING_NUM_ROWS, elasticsearch_date_to_pandas_date
 from eland import FieldMappings
 from eland import Index
 from eland import Operations
@@ -447,17 +447,17 @@ class QueryCompiler:
     def count(self):
         return self._operations.count(self)
 
-    def mean(self):
-        return self._operations.mean(self)
+    def mean(self, numeric_only=None):
+        return self._operations.mean(self, numeric_only=numeric_only)
 
-    def sum(self):
-        return self._operations.sum(self)
+    def sum(self, numeric_only=None):
+        return self._operations.sum(self, numeric_only=numeric_only)
 
-    def min(self):
-        return self._operations.min(self)
+    def min(self, numeric_only=None):
+        return self._operations.min(self, numeric_only=numeric_only)
 
-    def max(self):
-        return self._operations.max(self)
+    def max(self, numeric_only=None):
+        return self._operations.max(self, numeric_only=numeric_only)
 
     def nunique(self):
         return self._operations.nunique(self)
@@ -547,199 +547,6 @@ class QueryCompiler:
         aggregatable_field_name = self._mappings.aggregatable_field_name(display_name)
 
         return aggregatable_field_name
-
-
-def elasticsearch_date_to_pandas_date(value: Union[int, str], date_format: str) -> pd.Timestamp:
-    """
-    Given a specific Elasticsearch format for a date datatype, returns the
-    'partial' `to_datetime` function to parse a given value in that format
-
-    **Date Formats: https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-date-format.html#built-in-date-formats
-
-    Parameters
-    ----------
-    value: Union[int, str]
-        The date value.
-    date_format: str
-        The Elasticsearch date format (ex. 'epoch_millis', 'epoch_second', etc.)
-
-    Returns
-    -------
-    datetime: pd.Timestamp
-        From https://www.elastic.co/guide/en/elasticsearch/reference/current/date.html
-        Date formats can be customised, but if no format is specified then it uses the default:
-        "strict_date_optional_time||epoch_millis"
-        Therefore if no format is specified we assume either strict_date_optional_time
-        or epoch_millis.
-    """
-
-    if date_format is None:
-        try:
-            value = int(value)
-            return pd.to_datetime(value, unit='ms')
-        except ValueError:
-            return pd.to_datetime(value)
-    elif date_format == "epoch_millis":
-        return pd.to_datetime(value, unit='ms')
-    elif date_format == "epoch_second":
-        return pd.to_datetime(value, unit='s')
-    elif date_format == "strict_date_optional_time":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "basic_date":
-        return pd.to_datetime(value, format="%Y%m%d")
-    elif date_format == "basic_date_time":
-        return pd.to_datetime(value, format="%Y%m%dT%H%M%S.%f", exact=False)
-    elif date_format == "basic_date_time_no_millis":
-        return pd.to_datetime(value, format="%Y%m%dT%H%M%S%z")
-    elif date_format == "basic_ordinal_date":
-        return pd.to_datetime(value, format="%Y%j")
-    elif date_format == "basic_ordinal_date_time":
-        return pd.to_datetime(value, format="%Y%jT%H%M%S.%f%z", exact=False)
-    elif date_format == "basic_ordinal_date_time_no_millis":
-        return pd.to_datetime(value, format="%Y%jT%H%M%S%z")
-    elif date_format == "basic_time":
-        return pd.to_datetime(value, format="%H%M%S.%f%z", exact=False)
-    elif date_format == "basic_time_no_millis":
-        return pd.to_datetime(value, format="%H%M%S%z")
-    elif date_format == "basic_t_time":
-        return pd.to_datetime(value, format="T%H%M%S.%f%z", exact=False)
-    elif date_format == "basic_t_time_no_millis":
-        return pd.to_datetime(value, format="T%H%M%S%z")
-    elif date_format == "basic_week_date":
-        return pd.to_datetime(value, format="%GW%V%u")
-    elif date_format == "basic_week_date_time":
-        return pd.to_datetime(value, format="%GW%V%uT%H%M%S.%f%z", exact=False)
-    elif date_format == "basic_week_date_time_no_millis":
-        return pd.to_datetime(value, format="%GW%V%uT%H%M%S%z")
-    elif date_format == "strict_date":
-        return pd.to_datetime(value, format="%Y-%m-%d")
-    elif date_format == "date":
-        return pd.to_datetime(value, format="%Y-%m-%d")
-    elif date_format == "strict_date_hour":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H")
-    elif date_format == "date_hour":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H")
-    elif date_format == "strict_date_hour_minute":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M")
-    elif date_format == "date_hour_minute":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M")
-    elif date_format == "strict_date_hour_minute_second":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S")
-    elif date_format == "date_hour_minute_second":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S")
-    elif date_format == "strict_date_hour_minute_second_fraction":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f", exact=False)
-    elif date_format == "date_hour_minute_second_fraction":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f", exact=False)
-    elif date_format == "strict_date_hour_minute_second_millis":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f", exact=False)
-    elif date_format == "date_hour_minute_second_millis":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f", exact=False)
-    elif date_format == "strict_date_time":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "date_time":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "strict_date_time_no_millis":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S%z")
-    elif date_format == "date_time_no_millis":
-        return pd.to_datetime(value, format="%Y-%m-%dT%H:%M:%S%z")
-    elif date_format == "strict_hour":
-        return pd.to_datetime(value, format="%H")
-    elif date_format == "hour":
-        return pd.to_datetime(value, format="%H")
-    elif date_format == "strict_hour_minute":
-        return pd.to_datetime(value, format="%H:%M")
-    elif date_format == "hour_minute":
-        return pd.to_datetime(value, format="%H:%M")
-    elif date_format == "strict_hour_minute_second":
-        return pd.to_datetime(value, format="%H:%M:%S")
-    elif date_format == "hour_minute_second":
-        return pd.to_datetime(value, format="%H:%M:%S")
-    elif date_format == "strict_hour_minute_second_fraction":
-        return pd.to_datetime(value, format="%H:%M:%S.%f", exact=False)
-    elif date_format == "hour_minute_second_fraction":
-        return pd.to_datetime(value, format="%H:%M:%S.%f", exact=False)
-    elif date_format == "strict_hour_minute_second_millis":
-        return pd.to_datetime(value, format="%H:%M:%S.%f", exact=False)
-    elif date_format == "hour_minute_second_millis":
-        return pd.to_datetime(value, format="%H:%M:%S.%f", exact=False)
-    elif date_format == "strict_ordinal_date":
-        return pd.to_datetime(value, format="%Y-%j")
-    elif date_format == "ordinal_date":
-        return pd.to_datetime(value, format="%Y-%j")
-    elif date_format == "strict_ordinal_date_time":
-        return pd.to_datetime(value, format="%Y-%jT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "ordinal_date_time":
-        return pd.to_datetime(value, format="%Y-%jT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "strict_ordinal_date_time_no_millis":
-        return pd.to_datetime(value, format="%Y-%jT%H:%M:%S%z")
-    elif date_format == "ordinal_date_time_no_millis":
-        return pd.to_datetime(value, format="%Y-%jT%H:%M:%S%z")
-    elif date_format == "strict_time":
-        return pd.to_datetime(value, format="%H:%M:%S.%f%z", exact=False)
-    elif date_format == "time":
-        return pd.to_datetime(value, format="%H:%M:%S.%f%z", exact=False)
-    elif date_format == "strict_time_no_millis":
-        return pd.to_datetime(value, format="%H:%M:%S%z")
-    elif date_format == "time_no_millis":
-        return pd.to_datetime(value, format="%H:%M:%S%z")
-    elif date_format == "strict_t_time":
-        return pd.to_datetime(value, format="T%H:%M:%S.%f%z", exact=False)
-    elif date_format == "t_time":
-        return pd.to_datetime(value, format="T%H:%M:%S.%f%z", exact=False)
-    elif date_format == "strict_t_time_no_millis":
-        return pd.to_datetime(value, format="T%H:%M:%S%z")
-    elif date_format == "t_time_no_millis":
-        return pd.to_datetime(value, format="T%H:%M:%S%z")
-    elif date_format == "strict_week_date":
-        return pd.to_datetime(value, format="%G-W%V-%u")
-    elif date_format == "week_date":
-        return pd.to_datetime(value, format="%G-W%V-%u")
-    elif date_format == "strict_week_date_time":
-        return pd.to_datetime(value, format="%G-W%V-%uT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "week_date_time":
-        return pd.to_datetime(value, format="%G-W%V-%uT%H:%M:%S.%f%z", exact=False)
-    elif date_format == "strict_week_date_time_no_millis":
-        return pd.to_datetime(value, format="%G-W%V-%uT%H:%M:%S%z")
-    elif date_format == "week_date_time_no_millis":
-        return pd.to_datetime(value, format="%G-W%V-%uT%H:%M:%S%z")
-    elif date_format == "strict_weekyear" or date_format == "weekyear":
-        # TODO investigate if there is a way of converting this
-        raise NotImplementedError("strict_weekyear is not implemented due to support in pandas")
-        return pd.to_datetime(value, format="%G")
-        # Not supported in pandas
-        # ValueError: ISO year directive '%G' must be used with the ISO week directive '%V'
-        # and a weekday directive '%A', '%a', '%w', or '%u'.
-    elif date_format == "strict_weekyear_week" or date_format == "weekyear_week":
-        # TODO investigate if there is a way of converting this
-        raise NotImplementedError("strict_weekyear_week is not implemented due to support in pandas")
-        return pd.to_datetime(value, format="%G-W%V")
-        # Not supported in pandas
-        # ValueError: ISO year directive '%G' must be used with the ISO week directive '%V'
-        # and a weekday directive '%A', '%a', '%w', or '%u'.
-    elif date_format == "strict_weekyear_week_day":
-        return pd.to_datetime(value, format="%G-W%V-%u")
-    elif date_format == "weekyear_week_day":
-        return pd.to_datetime(value, format="%G-W%V-%u")
-    elif date_format == "strict_year":
-        return pd.to_datetime(value, format="%Y")
-    elif date_format == "year":
-        return pd.to_datetime(value, format="%Y")
-    elif date_format == "strict_year_month":
-        return pd.to_datetime(value, format="%Y-%m")
-    elif date_format == "year_month":
-        return pd.to_datetime(value, format="%Y-%m")
-    elif date_format == "strict_year_month_day":
-        return pd.to_datetime(value, format="%Y-%m-%d")
-    elif date_format == "year_month_day":
-        return pd.to_datetime(value, format="%Y-%m-%d")
-    else:
-        warnings.warn("The '{}' format is not explicitly supported."
-                      "Using pandas.to_datetime(value) to parse value".format(date_format),
-                      Warning)
-        # TODO investigate how we could generate this just once for a bulk read.
-        return pd.to_datetime(value)
-
 
 class FieldMappingCache:
     """
