+++ eland/tests/dataframe/test_datetime_pytest.py
@@ -96,12 +96,11 @@ class TestDataFrameDateTime(TestData):
         index_name = 'eland_test_generate_es_mappings'
 
         ed_df = ed.pandas_to_eland(df, ES_TEST_CLIENT, index_name, es_if_exists="replace", es_refresh=True)
-        ed_df_head = ed_df.head()
 
-        print(df.to_string())
-        print(ed_df.to_string())
-        print(ed_df.dtypes)
-        print(ed_df._to_pandas().dtypes)
+        #print(df.to_string())
+        #print(ed_df.to_string())
+        #print(ed_df.dtypes)
+        #print(ed_df._to_pandas().dtypes)
 
         assert_series_equal(df.dtypes, ed_df.dtypes)
