+++ eland/tests/dataframe/test_hist_pytest.py
@@ -39,7 +39,6 @@ class TestDataFrameHist(TestData):
             {'DistanceKilometers': pd_distancekilometers[0], 'FlightDelayMin': pd_flightdelaymin[0]})
 
         t = ed_flights[['DistanceKilometers', 'FlightDelayMin']]
-        print(t.columns)
 
         ed_bins, ed_weights = ed_flights[['DistanceKilometers', 'FlightDelayMin']]._hist(num_bins=num_bins)
